- some windows development - windows header includes
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@1049 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/av/StreamInfo.cpp
@@ -92,7 +92,7 @@ namespace org {
         return _s->codec->sample_fmt;
       }
 
-      int StreamInfo::getDuration() {
+      int64_t StreamInfo::getDuration() {
         return _s->duration;
       }
